reorg, fixups and wip on tests
