+++ frontend/src/app/api/auth/session/route.ts
@@ -1,9 +1,8 @@
-import { getSessionManager } from "src/services/auth/session";
+import { sessionManager } from "src/services/auth/session";
 
 import { NextResponse } from "next/server";
 
 export async function GET() {
-  const sessionManager = await getSessionManager();
   const currentSession = await sessionManager.getSession();
   if (currentSession) {
     return NextResponse.json({
