+++ frontend/tests/services/auth/session.test.ts
@@ -1,38 +1,12 @@
-import {
-  createSession,
-  decrypt,
-  deleteSession,
-  encrypt,
-  getSession,
-  getTokenFromCookie,
-} from "src/services/auth/session";
-
-type RecursiveObject = {
-  [key: string]: () => RecursiveObject | string;
-};
+import { SessionManager } from "src/services/auth/session";
 
 const getCookiesMock = jest.fn();
 const setCookiesMock = jest.fn();
 const deleteCookiesMock = jest.fn();
-const reallyFakeMockJWTConstructor = jest.fn();
-const setProtectedHeaderMock = jest.fn(() => fakeJWTInstance());
-const setIssuedAtMock = jest.fn(() => fakeJWTInstance());
-const setExpirationTimeMock = jest.fn(() => fakeJWTInstance());
-const signMock = jest.fn();
 const jwtVerifyMock = jest.fn();
 
-// close over the token
-// all of this rigmarole means that the mocked signing functionality will output the token passed into it
-const setJWTMocksWithToken = (token: string) => {
-  signMock.mockImplementation(() => token);
-};
-
-const fakeJWTInstance = (): RecursiveObject => ({
-  setProtectedHeader: setProtectedHeaderMock,
-  setIssuedAt: setIssuedAtMock,
-  setExpirationTime: setExpirationTimeMock,
-  sign: signMock,
-});
+const encodeTextMock = jest.fn((arg) => arg);
+const createPublicKeyMock = jest.fn((arg) => arg);
 
 const cookiesMock = () => {
   return {
@@ -46,92 +20,41 @@ jest.mock("next/headers", () => ({
   cookies: () => cookiesMock(),
 }));
 
-jest.mock("jose", () => ({
-  jwtVerify: (...args: unknown[]): unknown => jwtVerifyMock(...args),
-  SignJWT: function SignJWTMock(
-    this: {
-      setProtectedHeader: typeof jest.fn;
-      setIssuedAt: typeof jest.fn;
-      setExpirationTime: typeof jest.fn;
-      sign: typeof jest.fn;
-      token: string;
-    },
-    { token = "" } = {},
-  ) {
-    reallyFakeMockJWTConstructor();
-    setJWTMocksWithToken(token);
-    return {
-      ...fakeJWTInstance(),
-    };
-  },
-}));
-
 jest.mock("src/constants/environments", () => ({
   environment: {
     SESSION_SECRET: "session secret",
+    API_JWT_PUBLIC_KEY: "api secret",
   },
 }));
 
 jest.mock("src/utils/generalUtils", () => ({
-  encodeText: (arg: string): string => arg,
+  encodeText: (arg: string): string => encodeTextMock(arg),
 }));
 
-describe("encrypt", () => {
-  afterEach(() => jest.clearAllMocks());
-  it("calls all the JWT functions with expected values and returns expected value", async () => {
-    const token = "fakeToken";
-    const expiresAt = new Date();
-    const encrypted = await encrypt({ token, expiresAt });
+jest.mock("crypto", () => ({
+  createPublicKey: (arg: string): string => createPublicKeyMock(arg),
+}));
 
-    expect(reallyFakeMockJWTConstructor).toHaveBeenCalledTimes(1);
-
-    expect(setProtectedHeaderMock).toHaveBeenCalledTimes(1);
-    expect(setProtectedHeaderMock).toHaveBeenCalledWith({ alg: "HS256" });
-
-    expect(setIssuedAtMock).toHaveBeenCalledTimes(1);
-
-    expect(setExpirationTimeMock).toHaveBeenCalledTimes(1);
-    expect(setExpirationTimeMock).toHaveBeenCalledWith(expiresAt);
-
-    expect(signMock).toHaveBeenCalledTimes(1);
-    expect(signMock).toHaveBeenCalledWith("session secret");
-
-    // this is synthetic but generally proves things are working
-    expect(encrypted).toEqual(token);
+describe("SessionManager", () => {
+  let manager: SessionManager;
+  beforeEach(() => {
+    manager = new SessionManager();
   });
-});
-
-describe("decrypt", () => {
   afterEach(() => jest.clearAllMocks());
-  it("calls JWT verification with expected values and returns payload", async () => {
-    const cookie = "fakeCookie";
-    jwtVerifyMock.mockImplementation((...args) => ({ payload: args }));
-    const decrypted = await decrypt(cookie);
-
-    expect(jwtVerifyMock).toHaveBeenCalledTimes(1);
-    expect(jwtVerifyMock).toHaveBeenCalledWith(cookie, "session secret", {
-      algorithms: ["HS256"],
-    });
-
-    expect(decrypted).toEqual([
-      cookie,
-      "session secret",
-      { algorithms: ["HS256"] },
-    ]);
+  it("constructs correctly with necessary key values", () => {
+    expect(encodeTextMock).toHaveBeenCalledWith("session secret");
+    expect(createPublicKeyMock).toHaveBeenCalledWith("api secret");
+    expect(manager).toBeInstanceOf(SessionManager);
   });
 
-  it("returns null on error", async () => {
-    jwtVerifyMock.mockImplementation(() => {
-      throw new Error();
-    });
-    const cookie = "fakeCookie";
-    const decrypted = await decrypt(cookie);
-    expect(decrypted).toEqual(null);
-  });
-});
+  describe("getSession", () => {\
+    it('calls decrypt with the correct arguments and returns successfully', () => {
 
-describe("createSession", () => {
-  afterEach(() => jest.clearAllMocks());
+    })
+    it('returns null if client token decrypt does not return a payload and token', () => {})
+    it('returns null if api token decrypt does not return a payload', () => {})
+  });
+
   it("calls cookie.set with expected values", async () => {
     await createSession("nothingSpecial");
     expect(setCookiesMock).toHaveBeenCalledTimes(1);
@@ -142,15 +65,6 @@ describe("createSession", () => {
       sameSite: "lax",
       path: "/",
     });
-  });
-});
-
-describe("deleteSession", () => {
-  afterEach(() => jest.clearAllMocks());
-  it("calls cookie.delete with expected values", () => {
-    deleteSession();
-    expect(deleteCookiesMock).toHaveBeenCalledTimes(1);
-    expect(deleteCookiesMock).toHaveBeenCalledWith("session");
   });
 });
